refactor(operator_commands_dashboard): extract hosts API + helpers
Move all hosts CRUD + helpers (hosts_config_path, load_hosts, save_hosts,
is_local_host, host_entry_name, tag_local_membership, get_hosts, add_host,
remove_host) from routes.rs into a new sibling hosts.rs module.

Updates distributed.rs, tmux.rs, and tests to import host helpers from
the new location.

- routes.rs: 3384 -> 3219
- hosts.rs (NEW, 168)

Refs #1266
